Fix #107: Moved the stream logging out of the logging module
